# openapi.yaml
openapi: 3.0.1
info:
  title: Hello, World!
  description: '[MVP] ACME Company B2C API'
  termsOfService: http://acme.io/terms/
  contact:
    email: apiteam@acme.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.1.0
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
- url: https://localhost:8080/api/v1
- url: http://localhost:8080/api/v1
tags:
- name: store
  description: management of store profile
- name: store items
  description: management of store items - products or/and services
- name: store orders
  description: management of store orders
- name: store order payment
  description: management of store order's payment
- name: store customer
  description: managment of store customers
paths:
  /stores:
    get:
      tags:
      - store
      summary: List Stores
      operationId: listStore
      parameters:
      - name: nameTerm
        in: query
        schema:
          type: string
      - name: addresTerm
        in: query
        schema:
          type: string
      - name: pageSize
        in: query
        required: true
        schema:
          type: integer
          format: int64
          default: 20
      - name: pageIndex
        in: query
        required: true
        schema:
          type: integer
          format: int64
          default: 1
      - name: orderBy
        in: query
        description: |
          Syntax: [field\*Term only\*, fieldId] [asc, desc]. Example: field desc"
        required: true
        style: form
        explode: false
        schema:
          type: array
          items:
            type: string
      responses:
        200:
          description: List of stores
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StorePage'
        400:
          description: Store not created
          content:
            application/json:
              schema:
                type: string
                description: Error message
    post:
      tags:
      - store
      summary: Add new Store
      operationId: addStore
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Store'
        required: true
      responses:
        201:
          description: Store created
          content:
            application/json:
              schema:
                type: integer
                format: int64
        400:
          description: Store not created
          content:
            application/json:
              schema:
                type: string
                description: Error message
      x-codegen-request-body-name: body
  /stores/{storeId}:
    put:
      tags:
      - store
      summary: Update Store profile
      operationId: updateStore
      parameters:
      - name: storeId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Store'
        required: true
      responses:
        204:
          description: Store updated
          content: {}
        400:
          description: Store not created
          content:
            application/json:
              schema:
                type: string
                description: Error message
      x-codegen-request-body-name: body
  /stores/{storeId}/items:
    get:
      tags:
      - store items
      summary: list store items
      operationId: listStoreItems
      parameters:
      - name: storeId
        in: path
        required: true
        schema:
          type: integer
          format: int54
      - name: categoryTerm
        in: query
        required: true
        schema:
          type: string
          enum:
          - service
          - product
      - name: unitTerm
        in: query
        schema:
          type: string
      - name: nameTerm
        in: query
        schema:
          type: string
      - name: pageSize
        in: query
        required: true
        schema:
          type: integer
          format: int64
          default: 20
      - name: pageIndex
        in: query
        required: true
        schema:
          type: integer
          format: int64
          default: 1
      responses:
        200:
          description: list of items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreItemPage'
    post:
      tags:
      - store items
      summary: create store item - product/service
      operationId: addStoreItem
      parameters:
      - name: storeId
        in: path
        required: true
        schema:
          type: integer
          format: int54
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreItem'
        required: true
      responses:
        201:
          description: Store Item created
          content: {}
        400:
          description: Store not created
          content:
            application/json:
              schema:
                type: string
                description: Error message
      x-codegen-request-body-name: body
  /stores/{storeId}/orders:
    post:
      tags:
      - store orders
      summary: Create an store order for possible payment
      operationId: addStoreOrder
      parameters:
      - name: storeId
        in: path
        required: true
        schema:
          type: integer
          format: int54
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreOrder'
        required: true
      responses:
        201:
          description: Order created
          content:
            application/json:
              schema:
                type: integer
                format: int32
        400:
          description: Store not created
          content:
            application/json:
              schema:
                type: string
                description: Error message
      x-codegen-request-body-name: body
components:
  schemas:
    Store:
      type: object
      properties:
        storeId:
          type: integer
          format: int64
          example: 2023443219
        name:
          type: string
          example: ACME Peba Master ltda
        address:
          type: string
          example: rua bolinha peba, nº 100, uberlundia MG
    StorePage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Store'
        lastPage:
          type: boolean
          example: false
        pageSize:
          type: integer
          format: int32
          example: 20
        pageIndex:
          type: integer
          format: int32
          example: 1
    Customer:
      type: object
      properties:
        customerId:
          type: integer
          format: int64
        name:
          type: string
        address:
          type: string
    StoreItem:
      type: object
      properties:
        storeItemId:
          type: integer
          format: int64
          example: 10987645634
        name:
          type: string
          example: carne bolinha bovina
        unit:
          type: string
          example: kg
        image:
          type: string
          description: image hash
          example: AC98BFE120D
        category:
          type: string
          enum:
          - product
          - service
        lastPrice:
          type: number
          example: 20.99
    StoreItemPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/StoreItem'
        lastPage:
          type: boolean
          example: false
        pageSize:
          type: integer
          format: int32
          example: 20
        pageIndex:
          type: integer
          format: int32
          example: 1
    StoreItemPricing:
      type: object
      properties:
        storeItemPricingId:
          type: integer
          format: int64
        storeItemId:
          type: integer
          format: int64
        unitPrice:
          type: number
        createdAt:
          type: string
          format: date-time
    StoreOrderItem:
      type: object
      properties:
        storeItemPricingId:
          type: number
          example: 10.89
        quantity:
          type: number
          example: 0.12
        unit:
          type: string
          example: kg
    StoreOrder:
      type: object
      properties:
        storeOrderId:
          type: integer
          format: int64
        customerId:
          type: integer
          format: int64
        confirmationDate:
          type: string
          format: date-time
        status:
          type: string
          enum:
          - pending
          - accepted
          - rejected
        item:
          type: array
          items:
            $ref: '#/components/schemas/StoreOrderItem'
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
        petId:
          type: integer
          format: int64
        quantity:
          type: integer
          format: int32
        shipDate:
          type: string
          format: date-time
        status:
          type: string
          description: Order Status
          enum:
          - placed
          - approved
          - delivered
        complete:
          type: boolean
          default: false
      xml:
        name: Order
    Category:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
      xml:
        name: Category
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
        userStatus:
          type: integer
          description: User Status
          format: int32
      xml:
        name: User
    Tag:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
      xml:
        name: Tag
    Pet:
      required:
      - name
      - photoUrls
      type: object
      properties:
        id:
          type: integer
          format: int64
        category:
          $ref: '#/components/schemas/Category'
        name:
          type: string
          example: doggie
        photoUrls:
          type: array
          xml:
            name: photoUrl
            wrapped: true
          items:
            type: string
        tags:
          type: array
          xml:
            name: tag
            wrapped: true
          items:
            $ref: '#/components/schemas/Tag'
        status:
          type: string
          description: pet status in the store
          enum:
          - available
          - pending
          - sold
      xml:
        name: Pet
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
  parameters:
    pageSizeParam:
      name: pageSize
      in: query
      required: true
      schema:
        type: integer
        format: int64
        default: 20
    pageIndexParam:
      name: pageIndex
      in: query
      required: true
      schema:
        type: integer
        format: int64
        default: 1
    nameTerm:
      name: nameTerm
      in: query
      schema:
        type: string
    storeIdParam:
      name: storeId
      in: path
      required: true
      schema:
        type: integer
        format: int54
    addressTerm:
      name: addresTerm
      in: query
      schema:
        type: string
    orderByParam:
      name: orderBy
      in: query
      description: |
        Syntax: [field\*Term only\*, fieldId] [asc, desc]. Example: field desc"
      required: true
      style: form
      explode: false
      schema:
        type: array
        items:
          type: string
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: http://petstore.swagger.io/oauth/dialog
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header